+++ src/library-resource-definitions.ts
@@ -1,0 +1,131 @@
+export class Package {
+    static Filename = "package";
+}
+
+export class ThemeMapping {
+    static ChildKey = "mappings";
+}
+
+export class ResourceWithNamedAssetsBase {
+    static AssetsKey = "assets";
+}
+
+export class Categorization {
+    static GroupKey = "categorizations";
+    static Filename = "categorization";
+}
+
+export class ImageBackground {
+    static GroupKey = "assets/backgrounds/images";
+    static Filename = "background";
+}
+
+export class Clipart {
+    static GroupKey = "assets/elements/clipart";
+    static Filename = "clipart";
+}
+
+export class GridResource {
+    static GroupKey = "assets/elements/grids"
+    static Filename = "grid";
+}
+
+export class MaskBorder {
+    static GroupKey = "assets/borders/masks";
+    static Filename = "mask";
+}
+
+export class LineBorder {
+    static GroupKey = "assets/borders/lines";
+    static Filename = "line";
+}
+
+export class FrameBorder {
+    static GroupKey = "assets/borders/frames";
+    static Filename = "frame";
+}
+
+export class Theme {
+    static GroupKey = "themes";
+    static Filename = "theme";
+}
+
+export class FormatResource {
+    static GroupKey = "formats";
+    static Filename = "format";
+}
+
+export class Layout {
+    static GroupKey = "layouts";
+    static Filename = "layout";
+}
+
+export class Font {
+    static GroupKey = "fonts";
+    static Filename = "font";
+}
+
+export class StockPhoto {
+    static GroupKey = "photos";
+    static Filename = "photo";
+}
+
+export class Style {
+    static GroupKey = "styles";
+    static Filename = "style";
+}
+
+export class Product {
+    static GroupKey = "products";
+    static Filename = "product";
+}
+
+export class DesignerSettings {
+    static GroupKey = "designer-settings";
+    static Filename = "designer-settings";
+}
+
+export class RenderingSettingsResource {
+    static GroupKey = "rendering-settings";
+    static Filename = "rendering-settings";
+}
+
+export class RenderingPipelineResource {
+    static GroupKey = "rendering-pipelines";
+    static Filename = "rendering-pipeline";
+}
+
+export class Redirections {
+    static GroupKey = "redirections";
+    static Filename = "redirections";
+}
+
+export class EventsResource {
+    static GroupKey = "events";
+    static Filename = "events";
+}
+
+export class GeneratedImage {
+    static GroupKey = "assets/elements/generated-images";
+    static Filename = "generated-image";
+}
+
+export class TextGenerationSettings {
+    static GroupKey = "text-generation-settings";
+    static Filename = "text-generation-settings";
+}
+
+export class PrintFinishing {
+    static GroupKey = "print-finishings";
+    static Filename = "print-finishing";
+}
+
+export class VariableDataDefinition {
+    static GroupKey = "variable-data-definitions";
+    static Filename = "variable-data-definition";
+}
+
+export class Filter {
+    static GroupKey = "filters";
+    static Filename = "filter";
+}